+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Event channel size extracted into properties variable (`grab.channel-size`)
 
 ## [2.0.3] - 2020-01-28
 ### Changed